Prepare gene model from an object of TxDb

Generate an object of track for viewTracks by given parameters.

geneModelFromTxdb(
  txdb,
  orgDb,
  gr,
  chrom,
  start,
  end,
  strand = c("*", "+", "-"),
  txdump = NULL
)

Arguments

txdb
An object of TxDb

orgDb
An object of "OrgDb"

gr
An object of GRanges.

chrom
chromosome name, must be a seqname of txdb

start
start position

end
end position

strand
strand

txdump
output of as.list(txdb), a list of data frames that can be used to make the db again with no loss of information.

Value

An object of track

See also

Examples

library(TxDb.Hsapiens.UCSC.hg19.knownGene)

#> Loading required package: GenomicFeatures

#> Loading required package: AnnotationDbi

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

library(org.Hs.eg.db)

#> 

trs <- geneModelFromTxdb(TxDb.Hsapiens.UCSC.hg19.knownGene,
                         org.Hs.eg.db,
                         chrom="chr20", 
                         start=22560000, 
                         end=22565000, 
                         strand="-")