Class "trackStyle"

An object of class "trackStyle" represents track style.

An object of class "track" represents scores of a given track.

# S4 method for track
seqlevels(x)

# S4 method for track
seqlevelsStyle(x)

# S4 method for track
seqlevelsStyle(x) <- value

# S4 method for track
show(object)

# S4 method for track
$(x, name)

# S4 method for track
$(x, name) <- value

setTrackStyleParam(ts, attr, value)

# S4 method for track,character
setTrackStyleParam(ts, attr, value)

setTrackXscaleParam(ts, attr, value)

# S4 method for track,character
setTrackXscaleParam(ts, attr, value)

setTrackYaxisParam(ts, attr, value)

# S4 method for track,character
setTrackYaxisParam(ts, attr, value)

Arguments

x
an object of trackStyle

value
values to be assigned.

object
an object of trackStyle.

name
slot name of trackStyle

ts
An object of track.

attr
the name of slot of trackStyle object to be changed.

Details

The attr of setTrackXscaleParam could not only be a slot of xscale, but also be position. If the attr is set to position, value must be a list of x, y and label. For example setTrackXscaleParam(track, attr="position", value=list(x=122929675, y=4, label=500))

Slots

tracktype

"character" track type, could be peak or cluster. Default is "peak". "cluster" is not supported yet. For interaction data, it could be "heatmap" or "link".

color

"character" track color. If the track has dat and dat2 slot, it should have two values.

NAcolor

"character" NA color for interactionData.

breaks

"numeric" breaks for color keys of interactionData.

height

"numeric" track height. It should be a value between 0 and 1

marginTop

"numeric" track top margin

marginBottom

"numeric" track bottom margin

xscale

object of xscale, describe the details of x-scale

yaxis

object of yaxisStyle, describe the details of y-axis

ylim

"numeric" y-axis range

ylabpos

"character", ylable postion, ylabpos should be 'left', 'right', 'topleft', 'bottomleft', 'topright', 'bottomright', 'abovebaseline' or 'underbaseline'. For gene type track, it also could be 'upstream' or 'downstream'

ylablas

"numeric" y lable direction. It should be a integer 0-3. See par:las

ylabgp

A "list" object, It will convert to an object of class gpar. This is basically a list of graphical parameter settings of y-label.

dat

Object of class GRanges the scores of a given track. It should contain score metadata.

dat2

Object of class GRanges the scores of a given track. It should contain score metadata. When dat2 and dat is paired, dat will be drawn as positive value where dat2 will be drawn as negative value (-1 * score)

type

The type of track. It could be 'data', 'gene', 'transcript', 'scSeq', 'lollipopData' or 'interactionData'.

format

The format of the input. It could be "BED", "bedGraph", "WIG", "BigWig" or "BAM"

style

Object of class trackStyle

name

unused yet

See also

Examples

extdata <- system.file("extdata", package="trackViewer",
mustWork=TRUE)
fox2 <- importScore(file.path(extdata, "fox2.bed"), format="BED")
setTrackStyleParam(fox2, "color", c("red","green"))
setTrackXscaleParam(fox2, "gp", list(cex=.5))
setTrackYaxisParam(fox2, "gp", list(col="blue"))
fox2$dat <- GRanges(score=numeric(0))